+++ apps/canonical-bridge-ui/pages/mainnet/index.tsx
@@ -4,8 +4,6 @@ import {
   BridgeTransfer,
   BridgeRoutes,
 } from '@bnb-chain/canonical-bridge-widget';
-import { useMemo } from 'react';
-import { useAccount } from 'wagmi';
 
 import { useTransferConfig } from '@/token-config/mainnet/useTransferConfig';
 import { chains } from '@/token-config/mainnet/chains';
@@ -15,13 +13,6 @@ import { light } from '@/core/theme/light';
 import { useWalletModal } from '@/core/wallet/hooks/useWalletModal';
 import { WalletProvider } from '@/core/wallet/WalletProvider';
 import { Layout } from '@/core/components/Layout';
-
-const unsupportedChainsInWallets = [
-  {
-    wallets: ['BinanceW3WSDK', 'binanceWeb3Wallet'],
-    chains: [1101],
-  },
-];
 
 export const bridgeConfig: ICanonicalBridgeConfig = {
   assetPrefix: env.ASSET_PREFIX,
@@ -52,24 +43,11 @@ function BridgeWidget() {
   const transferConfig = useTransferConfig();
   const { onOpen } = useWalletModal();
 
-  const { connector } = useAccount();
-  const supportedChains = useMemo(() => {
-    return chains.filter((e) => {
-      if (connector) {
-        const unsupportedChains = unsupportedChainsInWallets.find((e) =>
-          e.wallets.includes(connector.id),
-        )?.chains;
-        return !unsupportedChains?.includes(e.id);
-      }
-      return true;
-    });
-  }, [connector]);
-
   return (
     <CanonicalBridgeProvider
       config={bridgeConfig}
       transferConfig={transferConfig}
-      chains={supportedChains}
+      chains={chains}
       onClickConnectWalletButton={onOpen}
     >
       <Layout>