+++ packages/canonical-bridge-widget/src/modules/aggregator/types.ts
@@ -198,3 +198,8 @@ export type AdapterType =
   | StargateAdapter
   | LayerZeroAdapter
   | MesonAdapter;
+
+export interface IFeeDetails {
+  value: string;
+  symbol: string;
+}