+++ packages/canonical-bridge-sdk/CHANGELOG.md
@@ -1,5 +1,38 @@
 # @bnb-chain/canonical-bridge-sdk
 
+## 0.4.6
+
+### Patch Changes
+
+- Use stargate & meson api to fetch chain & token config
+- 2538636: Remove duplicated tokens for stargate
+- c4f3ee7: Update token element's info
+- 7caf17e: Fix meson token display symbol issue
+
+## 0.4.6-alpha.3
+
+### Patch Changes
+
+- Fix meson token display symbol issue
+
+## 0.4.6-alpha.2
+
+### Patch Changes
+
+- Remove duplicated tokens for stargate
+
+## 0.4.6-alpha.1
+
+### Patch Changes
+
+- Update token element's info
+
+## 0.4.6-alpha.0
+
+### Patch Changes
+
+- Use stargate & meson api to fetch chain & token config
+
 ## 0.4.5
 
 ### Patch Changes